+++ apps/learn-card-app/src/components/learncard/MyLearnCardModal.types.ts
@@ -1,0 +1,5 @@
+// In a separate file so that Vite can hot-reload it without restarting the dev server
+export enum MyLearnCardModalViewModeEnum {
+    child = 'child',
+    guardian = 'guardian',
+}