+++ packages/learn-card-network/simple-signing-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @welibraryos/lca-api-client
 
+## 1.1.9
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-service@1.2.4
+
 ## 1.1.8
 
 ### Patch Changes