+++ packages/plugins/simple-signing-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lca-api-plugin
 
+## 1.1.9
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-client@1.1.9
+
 ## 1.1.8
 
 ### Patch Changes