+++ services/meta-mask-snap/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @learncard/meta-mask-snap
 
+## 1.0.103
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.4.7
+
 ## 1.0.102
 
 ### Patch Changes